Moved most code into a class, added documentation
Most parts of the previous code were refactored into a class,
properly documented and slightly improved. It is now possible
to compute (and print) SHA1 sums for each file.

The default random number generator was changed to 'legacy', so
that the (seeded) results should be identical for different NumPy
versions, without manually setting the RNG.